requirements.txt
gradio==4.28.3
matplotlib==3.7.4
numpy==1.24.1
Pillow==9.3.0
Pillow==10.3.0
rdkit==2023.9.4
Requests==2.31.0
torch==2.1.2+cu118
